let lzf support ByteBuffer

+++ src/main/java/com/ning/compress/lzf/ChunkDecoder.java
@@ -167,7 +167,7 @@ public abstract class ChunkDecoder
                 if (data[ptr] != LZFChunk.BYTE_Z || data[ptr+1] != LZFChunk.BYTE_V) {
                     throw new LZFException("Corrupt input data, block #"+blockNr+" (at offset "+ptr+"): did not start with 'ZV' signature bytes");
                 }
-                int type = (int) data[ptr+2];
+                int type = data[ptr+2];
                 int blockLen = uint16(data, ptr+3);
                 if (type == LZFChunk.BLOCK_TYPE_NON_COMPRESSED) { // uncompressed
                     ptr += 5;
